+++ install/update_94_95.php
@@ -44,6 +44,37 @@ function update94to95() {
    $migration->displayTitle(sprintf(__('Update to %s'), '9.5.0'));
    $migration->setVersion('9.5.0');
 
+   /** Encrypted FS support  */
+   if (!$DB->fieldExists("glpi_items_disks", "encryption_status")) {
+      $migration->addField("glpi_items_disks", "encryption_status", "integer", [
+            'after'  => "is_dynamic",
+            'value'  => 0
+         ]
+      );
+   }
+
+   if (!$DB->fieldExists("glpi_items_disks", "encryption_tool")) {
+      $migration->addField("glpi_items_disks", "encryption_tool", "string", [
+            'after'  => "encryption_status"
+         ]
+      );
+   }
+
+   if (!$DB->fieldExists("glpi_items_disks", "encryption_algorithm")) {
+      $migration->addField("glpi_items_disks", "encryption_algorithm", "string", [
+            'after'  => "encryption_tool"
+         ]
+      );
+   }
+
+   if (!$DB->fieldExists("glpi_items_disks", "encryption_type")) {
+      $migration->addField("glpi_items_disks", "encryption_type", "string", [
+            'after'  => "encryption_algorithm"
+         ]
+      );
+   }
+   /** /Encrypted FS support  */
+
    // ************ Keep it at the end **************
    $migration->executeMigration();